requirements.txt
grpcio==1.68.1
grpcio-health-checking==1.68.1
grpcio-tools==1.68.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
protobuf==5.29.1
pytest==8.3.4
pytest-asyncio==0.25.1
ruff==0.9.0
setuptools==75.6.0
xxhash==3.5.0
